Scrape versions from node mirror HTML
The node mirror is a giant HTML page of links. Scrape out the versions
and return them as a list in `fetch_version_list(...)`. Granted, I just
found an `index.json` file which might negate all this work, but this
seems worthy of a commit, considering I'm a Rust n00b and it took me ~1h
to get this far.