Use mirror JSON index instead of scraping
I started with scraping because that's what `fnm` does and I had no idea
how version managers worked and it was a nice excuse to read OCaml.
After looking closer, I see an `index.json` file with all the data I'm
trying to scrape. It's lovely, and it's on other mirrors too. Much
nicer!